connector: synchronize reads and writes
Ensure:
 - there is only one write being executed at any given time
 - there is only one read being executed at any given time
 - there is never both read and write being executed at the same time

This commit adds a clever mutex/queue mechanism that collects all
pending requests to save the data and executes only one write instead.

`connector.loadFromFile` bails out immediately when there is a pending
write.

The change in test/workspace fixes a unit-test that depended on the race
condition in the older version of the connector.