Fix order of keys in JSON files
Implement a new method `WorkspaceEntity.getConfigFromData` that converts
the Entity data from connector cache into an object that has properties
ordered in the same way as the Entity properties are ordered in the
Entity LDL definition.

Extend the semantics of the property flag `json`: when the value is
`false`, the property is not written to the JSON config file.

Use these two mechanism in all relevant places to ensure that
JSON config files contain only the relevant keys and the order of these
keys is always the same.

Add an internal `componentName` property to embedded Model* entities.
Implement a hook in WorkspaceEntity to automatically fill
the `componentName` property from the parent model.