test: isolate `testConnection` tests
Before this change, all `testConnection` tests shared the same set
of datasources. As a result, mysql tests were using a configuration
with a datasource with an invalid connector. The idea was to test
that the warning about the invalid connector is ignored unless it is
the connector of the tested datasource.

Recently, loopback or loopback-datasource-juggler fixed data-source
setup and throws an error when the connector is not valid. The tests
thus have to ensure the configuration is valid now, which is achieved
by cleaning up any extra datasources before each test.