Collect stderr from child instead of self
Make child process non-silent so that we can directly access the output
from its stdout and stderr.

Pipe child.stdout to process.stdout since that is equivalent to the
previous behaviour.

Accumulate child.stderr in an Array because that appears to be the
intent of the previous implementation.

Changes behaviour introduced in 8d8e85a