Workspace: do not forward HOST and PORT to child
Modify `Workspace.start` to remove `HOST` and `PORT` variables from
the environment passed to the spawned child process.

This fixes a problem in Studio, where `PORT=4000 strong-studio`
could not start the child process, because the child could not listen
on the port 4000, as that port was already in used by Studio itself.