Fix "npm test" to not run MySQL setup outside CI
The change introduced in #311 broke the behaviour of "npm test" when
running on a local developer's machine.

This commit adds a new switch "--ci-only" to "bin/setup-mysql" and
configures "npm test" to use this switch.

As a result, "npm test" does not trigger database setup when running
locally, but still triggers it when running on a CI server.